requirements.txt
PyAutoGUI==0.9.53
playsound==1.2.2
keyboard==0.13.5
python-dotenv==0.19.0
pywin32-ctypes==0.2.0
psutil==5.9.1
Pillow==9.1.0
opencv-python==4.5.3.56